# api/openapi.json
{
  "components": {
    "schemas": {
      "AppointmentCreateRequest": {
        "properties": {
          "barberName": {
            "maxLength": 100,
            "minLength": 0,
            "type": "string"
          },
          "bookingDate": {
            "format": "date",
            "type": "string"
          },
          "bookingTime": {
            "minLength": 1,
            "pattern": "^([01]\\d|2[0-3]):[0-5]\\d$",
            "type": "string"
          },
          "customerEmail": {
            "format": "email",
            "maxLength": 100,
            "minLength": 0,
            "type": "string"
          },
          "customerName": {
            "maxLength": 100,
            "minLength": 0,
            "type": "string"
          },
          "customerPhone": {
            "maxLength": 50,
            "minLength": 0,
            "type": "string"
          },
          "serviceType": {
            "maxLength": 100,
            "minLength": 0,
            "type": "string"
          }
        },
        "required": [
          "barberName",
          "bookingDate",
          "bookingTime",
          "customerEmail",
          "customerName",
          "customerPhone",
          "serviceType"
        ],
        "type": "object"
      },
      "AppointmentDashboardResponse": {
        "properties": {
          "page": {
            "$ref": "#/components/schemas/PageObject"
          },
          "stats": {
            "$ref": "#/components/schemas/AppointmentStats"
          }
        },
        "type": "object"
      },
      "AppointmentResponse": {
        "properties": {
          "barberName": {
            "type": "string"
          },
          "bookingDate": {
            "format": "date",
            "type": "string"
          },
          "bookingTime": {
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "customerEmail": {
            "type": "string"
          },
          "customerName": {
            "type": "string"
          },
          "customerPhone": {
            "type": "string"
          },
          "id": {
            "format": "int64",
            "type": "integer"
          },
          "publicId": {
            "type": "string"
          },
          "serviceType": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "type": "object"
      },
      "AppointmentStats": {
        "properties": {
          "approved": {
            "format": "int64",
            "type": "integer"
          },
          "approvedRevenue": {
            "format": "double",
            "type": "number"
          },
          "denied": {
            "format": "int64",
            "type": "integer"
          },
          "overdue": {
            "format": "int64",
            "type": "integer"
          },
          "pending": {
            "format": "int64",
            "type": "integer"
          },
          "progress": {
            "format": "int32",
            "type": "integer"
          },
          "total": {
            "format": "int64",
            "type": "integer"
          }
        },
        "type": "object"
      },
      "AppointmentUpdateRequest": {
        "properties": {
          "status": {
            "maxLength": 50,
            "minLength": 0,
            "pattern": "^(APPROVED|DENIED)$",
            "type": "string"
          }
        },
        "required": [
          "status"
        ],
        "type": "object"
      },
...
      "CancelRequest": {
        "properties": {
          "email": {
            "format": "email",
            "minLength": 1,
            "type": "string"
          }
        },
        "required": [
          "email"
        ],
        "type": "object"
      },
...
      "PageableObject": {
        "properties": {
          "offset": {
            "format": "int64",
            "type": "integer"
          },
          "paged": {
            "type": "boolean"
          },
          "pageNumber": {
            "format": "int32",
            "type": "integer"
          },
          "pageSize": {
            "format": "int32",
            "type": "integer"
          },
          "sort": {
            "$ref": "#/components/schemas/SortObject"
          },
          "unpaged": {
            "type": "boolean"
          }
        },
        "type": "object"
      },
...
      "PageObject": {
        "properties": {
          "content": {
            "items": {
              "$ref": "#/components/schemas/AppointmentResponse"
            },
            "type": "array"
          },
          "empty": {
            "type": "boolean"
          },
          "first": {
            "type": "boolean"
          },
          "last": {
            "type": "boolean"
          },
          "number": {
            "format": "int32",
            "type": "integer"
          },
          "numberOfElements": {
            "format": "int32",
            "type": "integer"
          },
          "pageable": {
            "$ref": "#/components/schemas/PageableObject"
          },
          "size": {
            "format": "int32",
            "type": "integer"
          },
          "sort": {
            "$ref": "#/components/schemas/SortObject"
          },
          "totalElements": {
            "format": "int64",
            "type": "integer"
          },
          "totalPages": {
            "format": "int32",
            "type": "integer"
          }
        },
        "type": "object"
      },
...
      "SortObject": {
        "properties": {
          "empty": {
            "type": "boolean"
          },
          "sorted": {
            "type": "boolean"
          },
          "unsorted": {
            "type": "boolean"
          }
        },
        "type": "object"
      }
    }
  },
  "info": {
    "title": "OpenAPI definition",
    "version": "v0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/appointments": {
      "get": {
        "description": "Can be filtered by status and customer name, with pagination.",
        "operationId": "getAllAppointments",
        "parameters": [
          {
            "description": "Filter by status (PENDING, APPROVED, DENIED)",
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Search customer name",
            "in": "query",
            "name": "search",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Page number (0-indexed)",
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 0,
              "format": "int32",
              "type": "integer"
            }
          },
          {
            "description": "Page size",
            "in": "query",
            "name": "size",
            "required": false,
            "schema": {
              "default": 10,
              "format": "int32",
              "maximum": 100,
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AppointmentDashboardResponse"
                }
              }
            },
            "description": "Successfully retrieved paginated appointments"
          },
          "400": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AppointmentDashboardResponse"
                }
              }
            },
            "description": "Invalid filter or pagination parameters"
          }
        },
        "summary": "View a list of scheduled appointments",
        "tags": [
          "Barber Appointment Management System"
        ]
      },
      "post": {
        "operationId": "createAppointment",
        "parameters": [
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AppointmentCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AppointmentResponse"
                }
              }
            },
            "description": "Successfully booked appointment"
          },
          "400": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AppointmentResponse"
                }
              }
            },
            "description": "Invalid request payload provided"
          }
        },
        "summary": "Create/Book a new appointment (Guest Access)",
        "tags": [
          "Barber Appointment Management System"
        ]
      }
    },
    "/api/v1/appointments/{id}": {
      "delete": {
        "operationId": "deleteAppointment",
        "parameters": [
          {
            "description": "The database ID of the appointment to delete",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successfully deleted appointment"
          },
          "404": {
            "description": "Appointment was not found"
          }
        },
        "summary": "Delete/Cancel an appointment",
        "tags": [
          "Barber Appointment Management System"
        ]
      },
      "get": {
        "operationId": "getAppointmentById",
        "parameters": [
          {
            "description": "The database ID of the appointment",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AppointmentResponse"
                }
              }
            },
            "description": "Successfully retrieved appointment"
          },
          "404": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AppointmentResponse"
                }
              }
            },
            "description": "Appointment was not found"
          }
        },
        "summary": "Get an appointment by its ID",
        "tags": [
          "Barber Appointment Management System"
        ]
      },
      "put": {
        "operationId": "updateAppointmentStatus",
        "parameters": [
          {
            "description": "The database ID of the appointment to update",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AppointmentUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AppointmentResponse"
                }
              }
            },
            "description": "Successfully updated appointment status"
          },
          "400": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AppointmentResponse"
                }
              }
            },
            "description": "Invalid request payload provided"
          },
          "404": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AppointmentResponse"
                }
              }
            },
            "description": "Appointment was not found"
          }
        },
        "summary": "Approve or Deny an appointment status (Owner Admin Access)",
        "tags": [
          "Barber Appointment Management System"
        ]
      }
    },
    "/api/v1/appointments/public/busy-slots": {
      "get": {
        "operationId": "getBusySlots",
        "parameters": [
          {
            "description": "Barber name",
            "in": "query",
            "name": "barberName",
            "required": true,
            "schema": {
              "maxLength": 100,
              "minLength": 0,
              "type": "string"
            }
          },
          {
            "description": "Booking date (YYYY-MM-DD)",
            "in": "query",
            "name": "bookingDate",
            "required": true,
            "schema": {
              "minLength": 1,
              "pattern": "^\\d{4}-\\d{2}-\\d{2}$",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "items": {
                    "type": "string"
                  },
                  "type": "array"
                }
              }
            },
            "description": "List of busy time slots returned"
          },
          "400": {
            "content": {
              "*/*": {
                "schema": {
                  "items": {
                    "type": "string"
                  },
                  "type": "array"
                }
              }
            },
            "description": "Invalid barber name or date format"
          }
        },
        "summary": "Get a list of busy time slots for a specific barber on a specific date (Guest Access)",
        "tags": [
          "Barber Appointment Management System"
        ]
      }
    },
    "/api/v1/appointments/public/cancel/{publicId}": {
      "put": {
        "operationId": "publicCancelAppointment",
        "parameters": [
          {
            "description": "The public UUID of the appointment",
            "in": "path",
            "name": "publicId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CancelRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Appointment successfully cancelled"
          },
          "400": {
            "description": "Invalid or mismatched email"
          },
          "404": {
            "description": "Appointment with given public ID not found"
          }
        },
        "summary": "Cancel an appointment securely as a guest by verifying the booking email (Guest Access)",
        "tags": [
          "Barber Appointment Management System"
        ]
      }
    },
...
    "/api/v1/customer/appointments/{publicId}": {
      "delete": {
        "operationId": "cancelMyAppointment",
        "parameters": [
          {
            "description": "Appointment public UUID",
            "in": "path",
            "name": "publicId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Appointment cancelled successfully"
          },
          "401": {
            "description": "Not authenticated"
          },
          "404": {
            "description": "Appointment not found or not owned by user"
          }
        },
        "summary": "Cancel my appointment",
        "tags": [
          "Customer Dashboard"
        ]
      }
    },
...
  },
  "servers": [
    {
      "description": "Generated server url",
      "url": "http://localhost:8080"
    }
  ],
...
}