# src/main/java/com/example/taskflow/appointment/AppointmentUpdateRequest.java
package com.example.taskflow.appointment;

import jakarta.validation.constraints.NotBlank;
import jakarta.validation.constraints.Pattern;
import jakarta.validation.constraints.Size;

public record AppointmentUpdateRequest(
    @NotBlank(message = "Status is required")
    @Size(max = 50)
    @Pattern(regexp = "^(APPROVED|DENIED)$", message = "Status must be APPROVED or DENIED")
    String status
) {}